Added Sensor feature, changed how to config with additional helper files, new status page (former sunrise page), now with arduinojson 6

+++ LED_wall/LEDMatrixEffects.h
@@ -68,12 +68,12 @@ void fire(long counter) {
   // COOLING: How much does the air cool as it rises?
   // Less cooling = taller flames.  More cooling = shorter flames.
   // Default 55, suggested range 20-100
-  uint8_t COOLING = 320/kMatrixHeight;
+  int COOLING = 320/kMatrixHeight;
 
   // SPARKING: What chance (out of 255) is there that a new spark will be lit?
   // Higher chance = more roaring fire.  Lower chance = more flickery fire.
   // Default 120, suggested range 50-200.
-  uint8_t SPARKING = 50;
+  int SPARKING = 50;
 
   for (int x = 0; x < kMatrixWidth; x++) {
     // Step 1.  Cool down every cell a little
